+++ src/mindlessgen/cli/cli_parser.py
@@ -92,6 +92,12 @@ def cli_parser(argv: Sequence[str] | None = None) -> dict:
         required=False,
         help="Scaling factor for van der Waals radii.",
     )
+    parser.add_argument(
+        "--scale-minimal-bondlength",
+        type=float,
+        required=False,
+        help="Minimum bond length scaling factor.",
+    )
 
     ### Molecule generation arguments ###
     parser.add_argument(
@@ -273,6 +279,7 @@ def cli_parser(argv: Sequence[str] | None = None) -> dict:
         "element_composition": args_dict["element_composition"],
         "forbidden_elements": args_dict["forbidden_elements"],
         "scale_vdw_radii": args_dict["scale_vdw_radii"],
+        "scale_minimal_bondlength": args_dict["scale_minimal_bondlength"],
     }
     # XTB specific arguments
     rev_args_dict["xtb"] = {"xtb_path": args_dict["xtb_path"]}